separating components for forecast and bug fixes
